Split security groups and rules from networks
