+++ os_migrate/plugins/modules/export_security_group.py
@@ -54,7 +54,7 @@ import openstack
 from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.os_migrate.os_migrate.plugins.module_utils import filesystem
-from ansible_collections.os_migrate.os_migrate.plugins.module_utils import network
+from ansible_collections.os_migrate.os_migrate.plugins.module_utils import security_group
 
 
 def run_module():
@@ -78,7 +78,7 @@ def run_module():
     conn = openstack.connect(cloud=module.params['cloud'])
     sdk_sec = conn.network.find_security_group(module.params['name'], ignoring_missing=False)
 
-    ser_sec = network.serialize_security_group(sdk_sec)
+    ser_sec = security_group.serialize_security_group(sdk_sec)
 
     result['changed'] = filesystem.write_or_replace_resource(
         module.params['path'], ser_sec)
